testutil: move int checkers to separate file
