+++ testutil/checkers_test.go
@@ -396,32 +396,6 @@ func (s *CheckersS) TestSystemCallSequenceEqual(c *check.C) {
 		[]CallResultError{}, []CallResultError{{C: `foo`}})
 }
 
-func (s *CheckersS) TestIntChecker(c *check.C) {
-	c.Assert(1, IntLessThan, 2)
-	c.Assert(1, IntLessEqual, 1)
-	c.Assert(1, IntEqual, 1)
-	c.Assert(2, IntNotEqual, 1)
-	c.Assert(2, IntGreaterThan, 1)
-	c.Assert(2, IntGreaterEqual, 2)
-
-	// Wrong argument types.
-	testCheck(c, IntLessThan, false, "left-hand-side argument must be an int", false, 1)
-	testCheck(c, IntLessThan, false, "right-hand-side argument must be an int", 1, false)
-
-	// Relationship error.
-	testCheck(c, IntLessThan, false, "relation 2 < 1 is not true", 2, 1)
-	testCheck(c, IntLessEqual, false, "relation 2 <= 1 is not true", 2, 1)
-	testCheck(c, IntEqual, false, "relation 2 == 1 is not true", 2, 1)
-	testCheck(c, IntNotEqual, false, "relation 2 != 2 is not true", 2, 2)
-	testCheck(c, IntGreaterThan, false, "relation 1 > 2 is not true", 1, 2)
-	testCheck(c, IntGreaterEqual, false, "relation 1 >= 2 is not true", 1, 2)
-
-	// Unexpected relation.
-	unexpected := &intChecker{CheckerInfo: &check.CheckerInfo{Name: "unexpected", Params: []string{"a", "b"}}, rel: "==="}
-	testCheck(c, unexpected, false, `unexpected relation "==="`, 1, 2)
-
-}
-
 func (s *CheckersS) TestPaddedChecker(c *check.C) {
 	type row struct {
 		lhs     string
